requirements.txt
# === Duma McDonald's - Sense Dashboard ===

# Web framework
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.8.2
jinja2==3.1.4

# Databricks SQL & Data Processing
databricks-sql-connector==3.3.0
pandas==2.2.2
numpy<2.0.0

# OpenAI (DUMA AI Agent)
openai==1.35.10
httpx==0.27.0

# Production Server (Linux/Docker)
gunicorn==22.0.0
